Fixed add-on mode, delete bug and set proper owner on videoversiontype

+++ jellyfin_kodi/objects/kodi/queries.py
@@ -825,6 +825,7 @@ WHERE           idFile = ?
 delete_video_version_type = """
 DELETE FROM     videoversiontype
 WHERE           id = ?
+AND             owner = 1
 """
 delete_set = """
 DELETE FROM     sets
